+++ SwiftNES/Implementation/Mappers/Mapper11.swift
@@ -9,15 +9,15 @@
 import Foundation
 
 final class Mapper11: Mapper {
-	private var prgBankOffset: UInt16
-	private var chrBankOffset: UInt16
+	private var prgBankOffset: Int
+	private var chrBankOffset: Int
 	
 	override init() {
 		prgBankOffset = 0
 		chrBankOffset = 0
 	}
 	
-	override func read(_ address: UInt16) -> UInt8 {
+	override func read(_ address: Int) -> UInt8 {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				return ppuMemory.banks[chrBankOffset + address]
@@ -32,7 +32,7 @@ final class Mapper11: Mapper {
 		return 0
 	}
 	
-	override func write(_ address: UInt16, data: UInt8) {
+	override func write(_ address: Int, data: UInt8) {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				ppuMemory.banks[chrBankOffset + address] = data
@@ -46,7 +46,7 @@ final class Mapper11: Mapper {
 	}
 	
 	func bankSelect(_ data: UInt8) {
-		prgBankOffset = (UInt16(data) & 0x3) * 0x8000
-		chrBankOffset = ((UInt16(data) & 0xF0) >> 4) * 0x2000
+		prgBankOffset = (Int(data) & 0x3) * 0x8000
+		chrBankOffset = ((Int(data) & 0xF0) >> 4) * 0x2000
 	}
 }
