+++ SwiftNES/Implementation/Mappers/Mapper2.swift
@@ -10,12 +10,12 @@ import Foundation
 
 final class Mapper2: Mapper {
 	
-	private var prgBank0Offset: UInt16
-	private var prgBank1Offset: UInt16
+	private var prgBank0Offset: Int
+	private var prgBank1Offset: Int
 	
 	override var cpuMemory: CPUMemory! {
 		didSet {
-			prgBank1Offset = UInt16(cpuMemory.banks.count - 0x4000 - 0xC000)
+			prgBank1Offset = Int(cpuMemory.banks.count - 0x4000 - 0xC000)
 		}
 	}
 	
@@ -24,7 +24,7 @@ final class Mapper2: Mapper {
 		prgBank1Offset = 0
 	}
 	
-	override func read(_ address: UInt16) -> UInt8 {
+	override func read(_ address: Int) -> UInt8 {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				return ppuMemory.banks[address]
@@ -42,7 +42,7 @@ final class Mapper2: Mapper {
 		return 0
 	}
 	
-	override func write(_ address: UInt16, data: UInt8) {
+	override func write(_ address: Int, data: UInt8) {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				ppuMemory.banks[address] = data
@@ -56,6 +56,6 @@ final class Mapper2: Mapper {
 	}
 	
 	func bankSelect(_ data: UInt8) {
-		prgBank0Offset = (UInt16(data) & 0xF) * 0x4000
+		prgBank0Offset = (Int(data) & 0xF) * 0x4000
 	}
 }
