+++ SwiftNES/Implementation/Mappers/Mapper3.swift
@@ -10,13 +10,13 @@ import Foundation
 
 final class Mapper3: Mapper {
 	
-	private var chrBankOffset: UInt16
+	private var chrBankOffset: Int
 	
 	override init() {
 		self.chrBankOffset = 0
 	}
 	
-	override func read(_ address: UInt16) -> UInt8 {
+	override func read(_ address: Int) -> UInt8 {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				return ppuMemory.banks[chrBankOffset + address]
@@ -31,7 +31,7 @@ final class Mapper3: Mapper {
 		return 0
 	}
 	
-	override func write(_ address: UInt16, data: UInt8) {
+	override func write(_ address: Int, data: UInt8) {
 		switch address {
 			case 0x0000 ..< 0x2000:
 				ppuMemory.banks[chrBankOffset + address] = data
@@ -45,6 +45,6 @@ final class Mapper3: Mapper {
 	}
 	
 	func bankSelect(_ data: UInt8) {
-		chrBankOffset = UInt16(data) * 0x2000
+		chrBankOffset = Int(data) * 0x2000
 	}
 }
